+++ test-neural/neural-wasm/mnist/src/lib.rs
@@ -42,7 +42,7 @@ impl MnistNetwork {
         Ok(MnistNetwork {
             network: model.network,
             accuracy: model.metadata.accuracy,
-            test_samples: model.metadata.test_samples,
+            test_samples: model.metadata.test_samples as usize,
             trained_at: model.metadata.trained_at,
             normalization: model.metadata.normalization,
         })
